Use String::Diff if installed for diagnostics
If String::Diff is installed it will be used when a test fails because
of a string comparison, This makes it easier to spot the differences
between what we wanted and what we got. If String::Diff is not installed
it will not attempt to use it.

Also added more Test::Exception tests