Made class for system fennec config
 * Refactored Fennec::Runner::init to properly use
   args/config/defaults/etc.
 * Divided init() into smaller pieces inside a Proto class

Fennec::Runner::init had enough logic that did not seem like it belonged
inside Fennec::Runner itself, but I also did not want to clutter the
Fennec.pm module with it. Creating a Proto class that builds the
prototype for the runner seemed a good idea.

We will see how this plays out, I have never seen this design before,
not sure if thats because its a bad idea. As well I used some clever
trickery to cache all the fields data, create accessors that return
(name => data), and have accessor based access to just the data for each
field.

In short: This Proto class may be too magical.